Keep workspace.last not_found when Focus Last stays in the workspace
Pane-scoped history can make the toggle land on another pane in the
current workspace; keep reporting not_found there so tmux - targets never
resolve to the current workspace. Also fix the test's JSONValue.int
argument type (Int64).

+++ cmuxTests/FocusHistoryScopeTests.swift
@@ -256,7 +256,7 @@ struct FocusHistoryScopeTests {
         #expect(manager.selectedTabId == third.id)
 
         let coordinator = ControlCommandCoordinator(context: TerminalController.shared)
-        func selectLast(_ id: Int) -> JSONValue? {
+        func selectLast(_ id: Int64) -> JSONValue? {
             let result = coordinator.handle(ControlRequest(id: .int(id), method: "workspace.last", params: [:]))
             settleFocusBroadcasts()
             guard case .ok(.object(let payload)) = result else { return nil }
